refactor: port to macos
event-manager is linux specific, it use to trigger fuse session,
http_thread exit. In this refactor impl, wait fuse channel
thread exit then exit the http_thread.

api-server use eventfd as the request notification to trigger
server run. Because of remove event-manger and eventfd, it can
not work. Add new thread named api-server for it do event
loop by self, and refactor api channel to Option.
If api-server receiver none request means for exit.

+++ rafs/src/metadata/direct_v6.rs
@@ -488,7 +488,7 @@ impl OndiskInodeWrapper {
 
     fn mode_format_bits(&self) -> u32 {
         let i = self.disk_inode();
-        i.mode() as u32 & libc::S_IFMT
+        i.mode() as u32 & libc::S_IFMT as u32
     }
 
     fn make_chunk_io(
@@ -884,17 +884,17 @@ impl RafsInode for OndiskInodeWrapper {
     }
 
     fn is_dir(&self) -> bool {
-        self.mode_format_bits() == libc::S_IFDIR
+        self.mode_format_bits() == libc::S_IFDIR as u32
     }
 
     /// Check whether the inode is a symlink.
     fn is_symlink(&self) -> bool {
-        self.mode_format_bits() == libc::S_IFLNK
+        self.mode_format_bits() == libc::S_IFLNK as u32
     }
 
     /// Check whether the inode is a regular file.
     fn is_reg(&self) -> bool {
-        self.mode_format_bits() == libc::S_IFREG
+        self.mode_format_bits() == libc::S_IFREG as u32
     }
 
     /// Check whether the inode is a hardlink.
